requirements.txt
colorama
pysocks
cryptography.fernet
tkinter